Standardize getDesign inputs (#48)
Nearly all getDesign methods now take a list of formulas as the key argument, instead of a mixture of double formulas, multiple arguments, or lists. The only exceptions are occuMulti and occuMS since they have so many more design matrices to generate. This allowed for removal or simplification of a bunch of methods.

In addition all unmarkedFit objects now have a formlist slot, and some unnecessary formula-related slots were removed.